+++ changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+- :feature:`-` allow specifying bump level at run time
 - :feature:`-` check (select) configuration keys for existence before proceeding
   with the rest of the script
 - :feature:`-` consolidate requirements to ``requirements.in``, and generate